Refactored to support new store / repository structure with multiple graphs.
