+++ src/extension/individual-node-provider.ts
@@ -1,6 +1,5 @@
 import * as vscode from 'vscode';
-import { DocumentContext } from '../mentor';
-import { ClassRepository, IndividualRepository } from '@faubulous/mentor-rdf';
+import { OntologyRepository } from '@faubulous/mentor-rdf';
 import { IndividualNode } from './individual-node';
 import { ResourceNodeProvider } from './resource-node-provider';
 import { ClassNode } from './class-node';
@@ -8,72 +7,61 @@ import { ClassNode } from './class-node';
 /**
  * A tree node provider for RDF properties.
  */
-export class IndividualNodeProvider extends ResourceNodeProvider<IndividualRepository> {
-
+export class IndividualNodeProvider extends ResourceNodeProvider<OntologyRepository> {
 	/**
 	 * Indicates whether the individual type should be shown as root nodes in the tree.
 	 */
 	showTypes: boolean = true;
 
 	/**
-	 * If defined, show only individuals of this type.
-	 */
-	typeFilter: string | undefined;
-
-	/**
-	 * A class repository that is used for creating individual type nodes.
-	 */
-	classRepository: ClassRepository | undefined;
-
-	/**
 	 * A set of URIs that are used for marking the nodes as classes for the getTreeItem method.
 	 */
 	classNodes: any = {};
 
-	override onDidChangeVocabularyContext(context: DocumentContext): void {
-		if (context?.store) {
-			this.repository = new IndividualRepository(context.store);
-			this.classRepository = new ClassRepository(context.store);
-		} else {
-			this.repository = undefined;
-			this.classRepository = undefined;
-		}
-	}
-
 	override getParent(uri: string): string | undefined {
-		return this.repository?.getIndividualTypes(uri).sort().slice(0, 1)[0];
+		if (this.context) {
+			return this.context.repository.getIndividualTypes(uri).sort().slice(0, 1)[0];
+		} else {
+			return undefined;
+		}
 	}
 
 	override getChildren(uri: string): string[] {
-		if (!this.repository) {
+		if (this.context) {
+			let result;
+
+			if (uri || !this.showTypes) {
+				result = this.context.repository.getIndividuals(this.context.graphs, uri).sort();
+			} else {
+				result = this.context.repository.getIndividualTypes(this.context.graphs).sort();
+
+				// Mark the nodes as classes for the getTreeItem method.
+				result.forEach((type: string) => this.classNodes[type] = true);
+			}
+
+			return result;
+		} else {
 			return [];
 		}
-
-		let result;
-
-		if (uri || !this.showTypes || this.typeFilter) {
-			const u = uri ?? this.typeFilter;
-
-			result = this.repository.getIndividuals(u).sort();
-		} else {
-			result = this.repository.getIndividualTypes().sort();
-
-			// Mark the nodes as classes for the getTreeItem method.
-			result.forEach((type: string) => this.classNodes[type] = true);
-		}
-
-		return result;
 	}
 
 	override getTreeItem(uri: string): vscode.TreeItem {
-		if (this.classNodes[uri]) {
-			return new ClassNode(this.classRepository!, uri, vscode.TreeItemCollapsibleState.Collapsed);
+		if (this.context) {
+			if (this.classNodes[uri]) {
+				return new ClassNode(this.context, uri, vscode.TreeItemCollapsibleState.Collapsed);
+			} else {
+				return new IndividualNode(this.context, uri);
+			}
 		} else {
-			return new IndividualNode(this.repository!, uri);
+			throw new Error('Invalid context.');
 		}
 	}
 
 	override getTotalItemCount(): number {
-		return this.repository ? this.repository.getIndividuals().length : 0;
+		if (this.context) {
+			return this.context.repository.getIndividuals(this.context.graphs).length;
+		} else {
+			return 0;
+		}
 	}
 }
