+++ src/extension/individual-node.ts
@@ -1,22 +1,23 @@
 import * as vscode from 'vscode';
-import { IndividualRepository } from '@faubulous/mentor-rdf';
+import { OntologyRepository } from '@faubulous/mentor-rdf';
+import { DocumentContext } from '../document-context';
 import { ResourceNode } from './resource-node';
 
-export class IndividualNode extends ResourceNode {
+export class IndividualNode extends ResourceNode<OntologyRepository> {
 	contextValue = 'individual';
 
 	type: 'individual' | 'class' = 'individual';
 
 	constructor(
-		protected readonly repository: IndividualRepository,
+		protected readonly context: DocumentContext<OntologyRepository>,
 		public readonly uri: string
 	) {
-		super(repository, uri);
+		super(context, uri);
 
 		this.collapsibleState = vscode.TreeItemCollapsibleState.None;
 
 		this.command = {
-			command: 'mentor.command.goToDefinition',
+			command: 'mentor.action.goToDefinition',
 			title: '',
 			arguments: [uri]
 		};
