+++ src/providers/index.ts
@@ -1,2 +1,7 @@
-export * from './rename-provider';
+export * from './completion-item-provider';
 export * from './definition-provider';
+export * from './feature-provider';
+export * from './hover-provider';
+export * from './reference-provider';
+export * from './rename-provider';
+export * from './semantic-tokens-provider';
